requirements.txt
gradio
numpy
opencv-python
pillow
torch==2.4.0
ultralytics
